backend/requirements.txt
# filepath: c:\Users\hvmar\Desktop\Block_mbt\backend\requirements.txt
alembic==1.16.5
annotated-types==0.7.0
anyio==4.10.0
asyncpg==0.30.0
bcrypt==4.0.1
beautifulsoup4==4.12.3
bleach==6.1.0
certifi==2025.10.5
cffi==2.0.0
click==8.2.1
colorama==0.4.6
cryptography==45.0.7
distro==1.9.0
dnspython==2.8.0
ecdsa==0.19.1
email-validator==2.3.0
fastapi==0.100.0
greenlet==3.2.4
h11==0.14.0
httpcore==0.16.3
httptools==0.6.4
httpx==0.23.3
httpx-oauth==0.10.1
idna==3.10
jiter==0.11.0
Mako==1.3.10
MarkupSafe==3.0.2
openai==2.3.0
passlib==1.7.4
pyasn1==0.6.1
pycparser==2.23
pydantic==1.10.23
pydantic_core==2.33.2
python-dotenv==1.1.1
python-http-client==3.3.7
python-jose==3.3.0
python-multipart==0.0.20
PyYAML==6.0.2
rsa==4.9.1
sendgrid==6.12.5
six==1.17.0
slowapi==0.1.9
sniffio==1.3.1
SQLAlchemy==1.4.41
sqlalchemy2-stubs==0.0.2a38
sqlmodel==0.0.8
starlette==0.27.0
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.15.0
uvicorn==0.22.0
watchfiles==1.1.0
websockets==15.0.1
Werkzeug==3.1.3